Add support for 32-bit platforms through a linker-level switch (#1571)
* Have stackalloc take a Size
* Move RawSize cast expr to its own logic
* Use correct intrinsic for UByte to UWord conversion
* Shift hardcoding of 64-bit words from NIR to the linker
* Eliminate hardcoding of Ptr/Word tags to 64-bit sizes and alignments
* Adjust Word/UWord API declarations to match other unsigned integer types
* Adjust runtime type information type definition to match struct
* Get initial code to run on i386
* Add i386 to build matrix
* Implement basic support for varargs
* Use is32 precomputed value instead of getting word size every time
* Fix conflicting types for unwinding APIs
* Temporarily check in ASAN changes to see why CI differs from local
* Specify arch for additional included tests
* Rename Word => Size
* Sync all gyb files and remove unsafe implicit from Int => USize
* Try to use compiler-rt
* Move is32 and sizeOfPtr to be part of linktime properties
* Fix boxing logic for sizes and set up ASAN build infra
* Enable Interflow optimizations for sizes
* Convert remaining size operations to non-intrinsic style
* Don't use intrinsic to convert ptr to size
* Fix memory bug in process_monitor
* Attempt to optimize memory usage from sizes
* Fix crash in no-opt tests
* Fix bug in allocating arrays of negative length
* Catch negative length early in InflaterInputStream
* Avoid USize allocations in Arrays to fix Immix crashes and update libunwind to fix optimized i386
* Switch to libunwind from LLVM 12 and include revision in vendored folder
* Use clang -v to get target triple
* Update docs
* Ignore file locking tests on 32-bit platforms
* Add specialized assumption for 32-bit non-compliant tests
* Reduce nesting in Arrays.scala
* Specialized assume for ASAN
* Add Scala 3 support after rebase
* Not binary compatible with 0.4.x
* Install Boehm into the cross-compiling image
* Include target platform when cross-compiling Boehm
* No emulator needed for i386
* Disable LTO on x86
* Fix compile failures in UdpSocketTest
* Fix NIR version, both major and minor should have been incremented

+++ nativelib/src/main/resources/scala-native/platform/unwind.h
@@ -8,8 +8,7 @@ int scalanative_unwind_init_local(void *cursor, void *context);
 int scalanative_unwind_step(void *cursor);
 int scalanative_unwind_get_proc_name(void *cursor, char *buffer, size_t length,
                                      void *offset);
-int scalanative_unwind_get_reg(void *cursor, int regnum,
-                               unsigned long long *valp);
+int scalanative_unwind_get_reg(void *cursor, int regnum, size_t *valp);
 int scalanative_unw_reg_ip();
 
 #endif
